CodeFactorの警告対策として2つ以上のクラスを定義しているcsファイルを分割する
CodeFactorの警告(保守性に難アリ)の対応で修正します。
