+++ tools/ChmSourceConverter/ChmSourceConverter/ChmSourceConverter.csproj
@@ -41,8 +41,11 @@
     <Reference Include="System.Data" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="EncoderEscapingFallbackBuffer.cs" />
     <Compile Include="EncoderEscapingFallback.cs" />
+    <Compile Include="LineEnumerator.cs" />
     <Compile Include="FileContents.cs" />
+    <Compile Include="ChmSourceConverterApp.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Settings.Designer.cs">
